news_board_API/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
click==7.1.2
Django==3.0.7
django-common-helpers==0.9.2
django-cron==0.5.1
djangorestframework==3.11.0
flake8==3.8.2
importlib-metadata==1.6.0
mccabe==0.6.1
pathspec==0.8.0
pkg-resources==0.0.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
regex==2020.5.14
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
zipp==3.1.0
